folder/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
channels==4.2.0
channels_redis==4.2.1
dj-database-url==2.3.0
Django==5.1.3
django-widget-tweaks==1.5.0
gunicorn==23.0.0
lz4==4.3.3
msgpack==1.1.0
packaging==24.2
psycopg2-binary==2.9.9
redis==5.2.0
sqlitecloud==0.0.83
sqlparse==0.5.2
typing_extensions==4.12.2
whitenoise==6.8.2
supabase~=1.0.3
